Reorganize tests and screenshot outputs for multiple workflow runs
