+++ docs/tutorials/single_session.rst
@@ -5,7 +5,7 @@ As a researcher, you’ve just completed an experimental session and you’d lik
 
 Upon launching the GUIDE, you'll begin on the Conversions page. If you’re opening the application for the first time, there should be no pipelines listed on this page.
 
-.. figure:: ./screenshots/home-page.png
+.. figure:: ../assets/tutorials/home-page.png
   :align: center
   :alt: Home page
 
@@ -28,14 +28,14 @@ The Project Setup page will have you define two pieces of information about your
 
 You’ll notice that the name property has a red asterisk next to it, which identifies it as a required property.
 
-.. figure:: ./screenshots/info-page.png
+.. figure:: ../assets/tutorials/single/info-page.png
   :align: center
   :alt: Project Setup page with no name (invalid)
 
 
 After specifying a unique project name, the colored background and error message will disappear, allowing you to advance to the next page.
 
-.. figure:: ./screenshots/valid-name.png
+.. figure:: ../assets/tutorials/single/valid-name.png
   :align: center
   :alt: Project Setup page with valid name
 
@@ -45,7 +45,7 @@ On this page, you’ll specify the type of **workflow** you’d like to follow f
 
 Since this is a single-session workflow, you’ll need to specify a **Subject ID** and **Session ID** to identify the data you’ll be converting.
 
-.. figure:: ./screenshots/workflow-page.png
+.. figure:: ../assets/tutorials/single/workflow-page.png
   :align: center
   :alt: Workflow page
 
@@ -55,33 +55,33 @@ Data Formats
 ^^^^^^^^^^^^
 Next, you’ll specify the data formats you’re working with on the Data Formats page. The GUIDE supports 40+ total neurophysiology formats. A full registry of available formats is available here.
 
-.. figure:: ./screenshots/formats-page.png
+.. figure:: ../assets/tutorials/single/formats-page.png
   :align: center
   :alt: Date Formats page
 
 The tutorial we're working with uses the SpikeGLX and Phy formats, a common output for NeuroPixel recordings and subsequent spike sorting. To specify that your pipeline will handle these files, you’ll press the “Add Format” button.
 
-.. figure:: ./screenshots/format-options.png
+.. figure:: ../assets/tutorials/single/format-options.png
   :align: center
   :alt: Format pop-up on the Data Formats page
 
 Then, select the relevant formats—in this case, **SpikeGLX Recording** and **Phy Sorting**—from the pop-up list. Use the search bar to filter for the format you need.
 
 
-.. figure:: ./screenshots/search-behavior.png
+.. figure:: ../assets/tutorials/single/search-behavior.png
   :align: center
   :alt: Searching for SpikeGLX in the format pop-up
 
 The selected formats will then display above the button.
 
 
-.. figure:: ./screenshots/interface-added.png
+.. figure:: ../assets/tutorials/single/interface-added.png
   :align: center
   :alt: Data Formats page with SpikeGLX Recording added to the list
 
 Advance to the next page when you have **SpikeGLX Recording** and **Phy Sorting** selected.
 
-.. figure:: ./screenshots/all-interfaces-added.png
+.. figure:: ../assets/tutorials/single/all-interfaces-added.png
   :align: center
   :alt: Data Formats page with both SpikeGLX Recording and Phy Sorting added to the list
 
@@ -92,7 +92,7 @@ Source Data Information
 ^^^^^^^^^^^^^^^^^^^^^^^
 On this page, specify the relevant **.bin** (Spikeglx) file and **phy** folder so that the GUIDE can find this source data to complete the conversion.
 
-.. figure:: ./screenshots/sourcedata-page-specified.png
+.. figure:: ../assets/tutorials/single/sourcedata-page-specified.png
   :align: center
   :alt: Source Data page with source locations specified
 
@@ -103,20 +103,30 @@ The file metadata page is a great opportunity to add rich annotations to the fil
 
 The General Metadata for the GUIDE has already been specified because the Session Start Time was automatically extracted from the SpikeGLX source data.
 
-.. figure:: ./screenshots/metadata-page.png
+.. figure:: ../assets/tutorials/single/metadata-nwbfile.png
   :align: center
   :alt: Metadata page with invalid Subject information
 
 
 However, we still need to add the Subject information—as noted by the red accents around that item. Let’s say that our subject is a male mouse with an age of P30D.
 
-.. figure:: ./screenshots/metadata-open.png
+.. figure:: ../assets/tutorials/single/metadata-subject-complete.png
   :align: center
   :alt: Metadata page with valid Subject information
 
   The status of the Subject information will update in real-time as you fill out the form.
 
-After you press the Next button, you’ll be presented with a progress bar that indicates how many files are left to run a stub conversion on.
+
+This dataset will also have **Ecephys** metadata extracted from the SpikeGLX source data.
+
+.. figure:: ../assets/tutorials/single/metadata-ecephys.png
+  :align: center
+  :alt: Ecephys metadata extracted from the SpikeGLX source data
+
+
+Let's leave this as-is and advance to the next page.
+
+As you advance, you’ll trigger a stub conversion and be shown a progress bar that indicates how many files are left.
 
 File Conversion
 ---------------
@@ -126,7 +136,7 @@ Inspector Report
 
 The Inspector Report page allows you to validate the generated stub NWB files against the latest Best Practices and determine whether you’ve provided all the metadata for your files. Using the internal sidebar, you can scope the displayed errors by a particular subject or session.
 
-.. figure:: ./screenshots/inspect-page.png
+.. figure:: ../assets/tutorials/single/inspect-page.png
   :align: center
   :alt: NWB Inspector report
 
@@ -138,7 +148,7 @@ Conversion Preview
 On the Conversion Preview, Neurosift allows you to explore the structure of the NWB file and ensure your data matches your expectations.
 
 
-.. figure:: ./screenshots/preview-page.png
+.. figure:: ../assets/tutorials/single/preview-page.png
   :align: center
   :alt: Neurosift preview visualization
 
@@ -149,7 +159,7 @@ Conversion Review
 
 Congratulations on finishing your first conversion of neurophysiology files using the NWB GUIDE!
 
-.. figure:: ./screenshots/conversion-results-page.png
+.. figure:: ../assets/tutorials/single/conversion-results-page.png
   :align: center
   :alt: Conversion results page with a list of converted files
