+++ tests/e2e/config.ts
@@ -1,0 +1,88 @@
+import { join, dirname } from 'node:path'
+import { fileURLToPath } from 'node:url'
+import { homedir } from 'node:os'
+import { existsSync } from 'node:fs'
+
+import paths from "../../paths.config.json" assert { type: "json" };
+import { connect } from '../puppeteer';
+
+// ------------------------------------------------------------------
+// ------------------------ Path Definitions ------------------------
+// ------------------------------------------------------------------
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+export const screenshotPath = join(__dirname, '..', '..', 'docs', 'assets', 'tutorials')
+const guideRootPath = join(homedir(), paths.root)
+const testRootPath = join(guideRootPath, '.test')
+export const testDataRootPath = join(testRootPath, 'test-data')
+const testDataPath = join(testDataRootPath, 'data')
+export const testDatasetPath = join(testDataRootPath, 'dataset')
+
+export const windowDims = {
+  width: 1280,
+  height: 800
+}
+
+export const alwaysDelete = [
+  join(testRootPath, 'pipelines'),
+  join(testRootPath, 'conversions'),
+  join(testRootPath, 'preview'),
+  join(testRootPath, 'config.json')
+]
+
+
+// -----------------------------------------------------------------------
+// ------------------------ Configuration Options ------------------------
+// -----------------------------------------------------------------------
+
+export const testInterfaceInfo = {
+  common: {
+    SpikeGLXRecordingInterface: {
+      id: 'SpikeGLX Recording',
+    },
+    PhySortingInterface: {
+      id: 'Phy Sorting'
+    }
+  },
+  multi: {
+    SpikeGLXRecordingInterface: {
+      format: '{subject_id}/{subject_id}_{session_id}/{subject_id}_{session_id}_g0/{subject_id}_{session_id}_g0_imec0/{subject_id}_{session_id}_g0_t0.imec0.ap.bin'
+    },
+    PhySortingInterface: {
+      format: '{subject_id}/{subject_id}_{session_id}/{subject_id}_{session_id}_phy'
+    }
+  },
+  single: {
+    SpikeGLXRecordingInterface: {
+      file_path: join(testDataPath, 'spikeglx', 'Session1_g0', 'Session1_g0_imec0', 'Session1_g0_t0.imec0.ap.bin')
+    },
+    PhySortingInterface: {
+      folder_path: join(testDataPath, 'phy')
+    }
+  }
+}
+
+export const subjectInfo = {
+  sex: 'M',
+  species: 'Mus musculus',
+  age: 'P30D'
+}
+
+// export const regenerateTestData = !existsSync(testDataRootPath) || false // Generate only if doesn't exist
+export const regenerateTestData = true // Force regeneration
+
+export const dandiInfo = {
+  id: '212750',
+  token: process.env.DANDI_STAGING_API_KEY
+}
+
+// -------------------------------------------------------
+// ------------------------ Tests ------------------------
+// -------------------------------------------------------
+
+export const publish = dandiInfo.token ? true : false
+
+if (!publish) console.log('No DANDI API key provided. Will skip dataset publication step...')
+
+
+export const references = connect()
